Another component state demo
Includes switching `DelegateFractionBasedBorderPainter` from `IntArray` to `LongArray` since `0xFFFFFFFF` is not a Kotlin integer

+++ theming/src/commonMain/kotlin/org/pushingpixels/aurora/theming/painter/border/DelegateFractionBasedBorderPainter.kt
@@ -33,7 +33,7 @@ import org.pushingpixels.aurora.theming.colorscheme.AuroraColorScheme
 class DelegateFractionBasedBorderPainter(
     override val displayName: String,
     val delegate: FractionBasedBorderPainter,
-    val masks: IntArray,
+    val masks: LongArray,
     val transform: (AuroraColorScheme) -> AuroraColorScheme
 ) : AuroraBorderPainter {
     override val isPaintingInnerOutline = false
